+++ changelog.md
@@ -1,3 +1,17 @@
+### 0.0.3
+#### example of Gustavo Barbieri
+* Mudanças em relação à versão anterior:
+    * classe GameObject: representa objetos do jogo (inimigos, tiros, jogador)
+    * classe Ship: base para todas as naves do jogo
+    * classe Enemy: presença de inimigos na tela
+
+
+### 0.0.2
+#### example of Gustavo Barbieri
+* Background: agora carrega uma imagem 
+e atualiza a posição durante o jogo, 
+dando a impressão que o fundo desce.
+
+
 ### 0.0.1 
-* install pygame
-* example of Gustavo Barbieri
+install pygame